na 15/09

+++ app/Http/Controllers/PostsController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StorePost;
 use App\Models\BlogPost;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 // use Illuminate\Support\Facades\DB;
 
 class PostsController extends Controller
@@ -48,7 +49,7 @@ class PostsController extends Controller
      */
     public function create()
     {
-        
+        // $this->authorize('posts.create');
 
         return view('posts.create');
     }
@@ -62,6 +63,7 @@ class PostsController extends Controller
     public function store(StorePost $request)
     {
         $validated = $request->validated();
+        $validated['user_id'] = $request->user()->id;
         $post = BlogPost::create($validated);
         // $post = new BlogPost();
         // $post->title = $validated['title'];
@@ -99,6 +101,13 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
+        $post = BlogPost::findOrFail($id);
+
+        $this->authorize($post);
+
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't edit this blog post");
+        // };
         return view('posts.edit', ['post' => BlogPost::findOrFail($id)]);
     }
 
@@ -112,6 +121,13 @@ class PostsController extends Controller
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        $this->authorize($post);
+
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't edit this blog post");
+        // };
+
         $validated = $request->validated();
         $post->fill($validated);
         $post->save();
@@ -130,6 +146,13 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+        $this->authorize($post);
+
+        // if (Gate::denies('delete-post', $post)) {
+        //     abort(403, "You can't delete this blog post");
+        // };
+
         $post->delete();
 
         session()->flash('status', 'Blog post was deleted!');
